# Significantly improves the performance when working with `PackageList`