@whyoleg 2025-08-20
> It seems we need some CacheService for all caches

Nice idea, create #4220!